requirements.txt
requests
spacy
email-validator
python-whois
phonenumbers
trafilatura[all]
ddgs
langchain
langchain-core
langchain-openai
langchain-community
docx2txt
pypdf
unstructured
